janked rfed side but blue farside works well
